add json scalar tests
